requirements.txt
appnope==0.1.3
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
executing==2.0.0
fastjsonschema==2.18.1
findspark==2.0.1
fonttools==4.43.1
idna==3.4
import-ipynb==0.1.4
ipynb==0.5.1
ipython==8.16.1
jedi==0.19.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_core==5.3.2
kiwisolver==1.4.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
mysql-connector==2.2.9
nbformat==5.9.2
numpy==1.26.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
progress==1.6
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
Pygments==2.16.1
PyMySQL==1.1.0
pyparsing==3.1.1
pyspark==3.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
seaborn==0.13.0
six==1.16.0
stack-data==0.6.3
tqdm==4.66.1
traitlets==5.11.2
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
